Create Delete the Middle Node of a Linked List.cpp
You are given the head of a linked list. Delete the middle node, and return the head of the modified linked list.

The middle node of a linked list of size n is the ⌊n / 2⌋th node from the start using 0-based indexing, where ⌊x⌋ denotes the largest integer less than or equal to x.

For n = 1, 2, 3, 4, and 5, the middle nodes are 0, 1, 1, 2, and 2, respectively.
////////////////////////////////////////////////////////////////////

Solution:
 Runtime: 1728 ms, faster than 35.73% of C++ online submissions for Delete the Middle Node of a Linked List.
Memory Usage: 294.8 MB, less than 21.93% of C++ online submissions for Delete the Middle Node of a Linked List.